Fix - Integer to Double Tax Percentage for getByName

+++ src/main/java/com/pos_main/Dao/TaxDao.java
@@ -12,7 +12,7 @@ public interface TaxDao {
 	
     TaxDto checkTaxAvailability(Integer id);
 	
-	List<TaxDto> getTaxByName(Integer taxPercentage);
+	List<TaxDto> getTaxByName(Double taxPercentage);
 	
     List<TaxDto> getAll();
